Remove HiForest mangling of Pat sequences
It turns out that one of the photon Pat sequences (patPhotonIsolation) was removed at some point, but not all references to it were cleanly removed from CMSSW. Originally we had to patch a single file in the PatAlgos directory to fix this, but that pulls in that directory for compilation which is messy.

This patch works around this issue by make a new local version of the patSequence instead of using the older bad file.